restructured folders to solve crashes
