--- trial_imports.py
@@ -1,8 +1,0 @@
-
-import sys
-import os
-import pathlib
-
-sys.path.append(os.path.join(pathlib.Path(__file__).parent, "my_utils"))
-print(os.path.join(pathlib.Path(__file__).parent, "my_utils"))
-# print(sys)
